socket refactor
