+++ frontend/syntaxmeets/src/services/socket.js
@@ -1,5 +1,5 @@
 import io from "socket.io-client";
 
 export const socket = io.connect(
-    process.env.REACT_APP_SYNTAXMEETS_BACKEND_API,{ transport: ["websocket"] }
+    process.env.REACT_APP_SYNTAXMEETS_BACKEND_API, { transport: ["websocket"] }
 );
